backend/requirements.txt
# NASA Space Apps Challenge 2025 - "Will It Rain on My Parade?"
flask
flask-cors
requests
python-dotenv
pandas
numpy